Moved Home widget to a separate widget
